+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Version 0.9.16 (2026-05-)
 * Log version on startup.
+* Start using Go testing package.
 * [Alerts/popups?]
 * [ChangeLog for users on new versions?]